+++ src/lib/components/json/JsonReadOnly.tsx
@@ -3,6 +3,7 @@ import dynamic from "next/dynamic";
 import { useMemo, useState } from "react";
 
 import { CopyButton } from "../copy";
+import { AmpTrackExpand } from "lib/services/amplitude";
 import { jsonLineCount, jsonValidate } from "lib/utils";
 
 import { ViewFullMsgButton } from "./ViewFullMsgButton";
@@ -17,6 +18,7 @@ interface JsonReadOnlyProps {
   canCopy?: boolean;
   isExpandable?: boolean;
   fullWidth?: boolean;
+  amptrackSection?: string;
 }
 
 const THRESHOLD_LINES = 16;
@@ -27,6 +29,7 @@ const JsonReadOnly = ({
   canCopy,
   isExpandable,
   fullWidth,
+  amptrackSection,
 }: JsonReadOnlyProps) => {
   const [viewFull, setViewFull] = useState(false);
 
@@ -82,7 +85,14 @@ const JsonReadOnly = ({
       )}
       {showExpandButton && (
         <ViewFullMsgButton
-          onClick={() => setViewFull((prev) => !prev)}
+          onClick={() => {
+            AmpTrackExpand(
+              viewFull ? "collapse" : "expand",
+              "json",
+              amptrackSection
+            );
+            setViewFull((prev) => !prev);
+          }}
           viewFull={viewFull}
         />
       )}
@@ -93,7 +103,7 @@ const JsonReadOnly = ({
           right="10px"
           className="copy-button-box"
         >
-          <CopyButton value={text} />
+          <CopyButton value={text} amptrackSection={amptrackSection} />
         </Box>
       )}
     </Box>