+++ src/lib/services/amplitude.tsx
@@ -1,6 +1,7 @@
 import { track } from "@amplitude/analytics-browser";
 
 import type { AttachFundsType } from "lib/components/fund/types";
+import type { Option } from "lib/types";
 
 export enum AmpEvent {
   INVALID_STATE = "To Invalid State",
@@ -82,6 +83,10 @@ export enum AmpEvent {
   USE_QUICK_EDIT_CODE = "Use Quick Edit Code",
   USE_OTHER_MODAL = "Use Other Modal",
   USE_SUBMIT_PROJECT = "Use Submit Project",
+  USE_VIEW_JSON = "Use View Json",
+  USE_UNSUPPORTED_ASSETS = "Use Unsupported Assets",
+  USE_TX_MSG_EXPAND = "Use Transaction Message Expand",
+  USE_EXPAND = "Use General Expand",
   // TX
   TX_SUCCEED = "Tx Succeed",
   TX_FAILED = "Tx Failed",
@@ -116,8 +121,10 @@ export const AmpTrackInvalidState = (title: string) =>
   track(AmpEvent.INVALID_STATE, { title });
 
 export const AmpTrack = (
-  event: Exclude<AmpEvent, ActionAmpEvent | SpecialAmpEvent>
-) => track(event);
+  event: Exclude<AmpEvent, ActionAmpEvent | SpecialAmpEvent>,
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  properties?: Record<string, any>
+) => track(event, properties);
 
 export const AmpTrackAction = (
   event: ActionAmpEvent,
@@ -156,3 +163,18 @@ export const AmpTrackSocial = (url: string) => track(AmpEvent.SOCIAL, { url });
 
 export const AmpTrackCelatone = (url: string) =>
   track(AmpEvent.CELATONE, { url });
+
+export const AmpTrackViewJson = (page: string) =>
+  track(AmpEvent.USE_VIEW_JSON, { page });
+
+export const AmpTrackUnsupportedToken = (page: Option<string>) =>
+  track(AmpEvent.USE_UNSUPPORTED_ASSETS, { page });
+
+export const AmpTrackCopier = (section: Option<string>, type: string) =>
+  track(AmpEvent.USE_COPIER, { section, type });
+
+export const AmpTrackExpand = (
+  action: "expand" | "collapse",
+  component: "assets" | "json" | "permission_address" | "event_box",
+  section: Option<string>
+) => track(AmpEvent.USE_EXPAND, { action, component, section });